backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
SQLAlchemy==2.0.40
pymysql==1.1.1
python-dotenv==1.1.0
loguru==0.7.3
openai==1.75.0
httpx==0.28.1
passlib==1.7.4
python-jose==3.4.0
bcrypt==4.3.0
pytest==8.3.5
pytest-asyncio==0.26.0
alembic==1.15.2 